Phase 1: Domain models, ports, and LargeTransferFilter (#46)
* Add domain model records with test fixtures and unit tests (#7)

Implements SOL-7: 8 domain model records (SolanaTransaction, Account,
LargeTransfer, Memo, FailedTransaction, Slot, BatchResult, IndexerStats)
with @Builder(toBuilder=true). Zero framework imports — only Lombok + java.*.
Includes test fixtures and builder/toBuilder unit tests for all models.


* Add domain port interfaces for hexagonal architecture (#8)

Implements SOL-8: 8 port interfaces (TransactionStream, TransactionRepository,
FailedTransactionRepository, TransferRepository, MemoRepository,
AccountRepository, StatsRepository, MetricsRecorder). Plain interfaces with
no annotations — only domain model types in method signatures.


* Add LargeTransferFilter pure function for >1.0 SOL detection (#9)

Implements SOL-9: static filter with LARGE_TRANSFER_THRESHOLD_SOL constant.
Strictly greater-than comparison — exactly 1.0 SOL is not a large transfer.
Includes 3 test cases covering large, small, and exact-threshold amounts.


* Use UUID-based signatures in fixture builders per testing standards

Builder factories now generate unique signatures/pubkeys on each call
using UUID, matching the pattern in TESTING_STANDARDS.md. Fixes
shouldAllowNullableFields to use SOME_TRANSACTION.toBuilder() instead
of calling transactionBuilder() twice (which would produce mismatched
signatures with UUID).


* Refactor LargeTransferFilter tests to use @ParameterizedTest

Replaces loop-based assertions with @ParameterizedTest + @ValueSource
so each amount value runs as a separate test case with better failure
reporting. Addresses CodeRabbit review feedback on PR #46.


* Use BigDecimal for SOL amounts instead of double

Monetary values should not use floating-point arithmetic. Changes
double to BigDecimal in SolanaTransaction.amount, LargeTransfer.amount,
LargeTransferFilter threshold, and TransferRepository.findByMinAmount.
Uses string constructor (new BigDecimal("1.0")) to avoid precision loss
and compareTo() for threshold comparison.


* Add DDD value objects, constructor validation, and aggregate projections

Introduces Signature and Pubkey as strongly-typed value objects with
compact constructor validation (non-null, non-blank, max length).
All domain records now validate invariants: requireNonNull for reference
types, non-negative checks for primitives.

SolanaTransaction becomes the aggregate root with toLargeTransfer(),
toMemo(), and toFailedTransaction() projection methods. Port interfaces
updated to accept Signature/Pubkey instead of raw strings.


* Update standards docs with DDD patterns and BigDecimal conventions

CLAUDE.md: add value objects, BigDecimal, compact constructor validation
to style rules, design decisions, and pre-commit checklist.

CODING_STANDARDS.md: add sections for value objects (Signature, Pubkey),
aggregate root projections, compact constructor validation rules, and
BigDecimal for monetary amounts. Update port interface examples.

TESTING_STANDARDS.md: add validation assertion patterns, aggregate
projection test patterns, update fixture examples with value objects
and BigDecimal.


* Update implementation plan, issues, and migration for DDD changes

Sync all docs with current domain model: Signature/Pubkey value objects,
BigDecimal amounts, aggregate root projections, compact constructor
validation. Update port interface signatures in issue definitions.
Change large_transfers.amount from DOUBLE PRECISION to NUMERIC in
migration and functional spec to match BigDecimal in Java.


* Add DDD classification guide to coding standards

Documents the decision framework for choosing between value objects,
aggregate roots, and projections. Includes decision tree, identification
criteria, implementation rules, and when to wrap vs not wrap primitives.


---------

+++ prism/src/main/resources/db/migration/V1__create_tables.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS large_transfers (
     id         SERIAL PRIMARY KEY,
     signature  VARCHAR(88) NOT NULL,
     slot       BIGINT NOT NULL,
-    amount     DOUBLE PRECISION NOT NULL,
+    amount     NUMERIC NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
